Move ReadOnlyTables to the table subpackage
