+++ src/main/java/seedu/address/storage/JsonTablesStorage.java
@@ -13,7 +13,7 @@ import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.util.JsonUtil;
-import seedu.address.model.ReadOnlyTables;
+import seedu.address.model.table.ReadOnlyTables;
 
 /**
  * A class to access RestOrRant data stored as a json file on the hard disk.
